+++ problem.h
@@ -17,10 +17,10 @@ private:
     // almacenamiento datos
     int numClusters;
     vector<int> limClusters;
-    vector<vector<double>> dataset;
+    vector<vector<long double>> dataset;
     int points;
     int variables;
-    vector<vector<vector<double>>> coordinates;
+    vector<vector<vector<long double>>> coordinates;
     // metodos auxiliares
     void loadClusters();
     void loadDataset();
@@ -33,26 +33,26 @@ public:
     // metodos almacenamiento
     int getNumClusters() const;
     const vector<int>& getLimClusters() const;
-    const vector<vector<double>>& getDataset() const;
+    const vector<vector<long double>>& getDataset() const;
     int getPoints() const;
     int getVariables() const;
-    const vector<vector<vector<double>>>& getClustersCoordinates() const;
+    const vector<vector<vector<long double>>>& getClustersCoordinates() const;
     // metodos impresion
     void printVector1D(const vector<int>& vector) const;
     void printVector2D(const vector<vector<double>>& vector) const;
     // metodo carga dataset
     void loadData();
     // metodos de evaluacion
-    vector<vector<vector<double>>> calculatePointsCoordinatesPerCluster(const vector<int>& assignment) const;
-    vector<double> calculateClusterValues(const vector<vector<vector<double>>>& coordinates) const;
-    double calculateFitness(const vector<double>& clusterValues) const;
+    vector<vector<vector<long double>>> calculatePointsCoordinatesPerCluster(const vector<int>& assignment) const;
+    vector<long double> calculateClusterValues(const vector<vector<vector<long double>>>& coordinates) const;
+    long double calculateFitness(const vector<long double>& clusterValues) const;
     // evaluacion completa
-    double evaluateSolution(const vector<int>& assignment,
-                            vector<vector<vector<double>>>& coordinates,
-                            vector<double>& clusterValues) const;
+    long double evaluateSolution(const vector<int>& assignment,
+                            vector<vector<vector<long double>>>& coordinates,
+                            vector<long double>& clusterValues) const;
     // metodo distancia entre vectores de soluciones de clusters
-    double distance(const vector<vector<vector<double>>>& vectorA,
-                    const vector<vector<vector<double>>>& vectorB) const;
+    long double distance(const vector<vector<vector<long double>>>& vectorA,
+                    const vector<vector<vector<long double>>>& vectorB) const;
 };
 
 #endif
